Updated tsconfig to work with file restructure
